refactor: move computeCowInventoryForNextDay to its own file
